feat(functional_handlers.go): add matchTableName and onTablesReqFeedbackSuccess functions
The matchTableName function has been added to compare the name of a
table with a given string. This function will be useful in finding a
specific table by its name. The onTablesReqFeedbackSuccess function has
been added to handle successful requests for table data. It uses the
matchTableName function to find a specific table and returns a
fieldsReqFeedback object with the ID of the found table. This will allow
for more efficient handling of successful table data requests.