feat(action.go): add allPhenoTableIds function to map table names to ids
The function allPhenoTableIds has been added to map table names to their
respective ids. This function will be useful in scenarios where we need
to quickly retrieve the id of a table using its name, improving the
efficiency of the code. It takes a PhenotypeTableManager instance and a
cli context as arguments, and returns a map of table names to ids.