feat(main.go): add create-strain-table command to create a table with strain annotation fields
The create-strain-table command has been added to allow users to create
a table with strain annotation fields preset. This feature enhances the
functionality of the application by providing more options for table
creation. The usage description of the create-phenotype-table command
has been updated to improve clarity and understanding of its function.