feat: add dagger.json for modware-import with golang dependency
The dagger.json file has been added to the modware-import module. This
file specifies the dependencies of the module, in this case, golang.
This change is necessary to ensure that the module has all the necessary
dependencies to function correctly. The dagger.json file also specifies
the engine version to ensure compatibility.