refactor(reader.go): split NewReader into NewReader and NewReaderFromStream for better flexibility and reusability
The NewReader function was split into two functions: NewReader and
NewReaderFromStream. This change allows the code to be more flexible and
reusable. Now, it's possible to create a new reader from a file (using
NewReader) or from any io.Reader (using NewReaderFromStream), which can
be a file, a network connection, or any other source of data. This makes
the code more versatile and easier to test, as we can now pass a buffer
of bytes to NewReaderFromStream in our tests.