test(reader_test.go): modify test cases to include a new parameter in NewReader function
The NewReader function now includes a new parameter to decide whether to
skip the first row or not. This change is reflected in all the test
cases where NewReader function is used. A new test case is added to
verify the functionality of skipping the first row when the new
parameter is set to true.