feat(load.go): implement concurrent execution of loader functions
The changes introduce a new function `executeLoaderSlice` that runs a
slice of loader functions concurrently using goroutines. This is done to
improve the performance of the application by allowing multiple loader
functions to execute simultaneously. The function also handles errors
and results from the loader functions, cancelling all other functions if
an error occurs in any one of them. The `processFnRunnerProperties`
function is also added to process the results and errors from the
`executeLoaderSlice` function.