feat(main.go, flag.go, load.go): add load-strain-annotation command to load strain data from excel to baserow table
The new command "load-strain-annotation" is added to load strain
annotation data from an excel spreadsheet to a baserow table. This
feature enhances the functionality of the application by allowing users
to import strain data directly from an excel file. The command requires
a database token with write privilege, a database table id, an input
excel spreadsheet file with strain annotations, and optionally, the name
of the sheet which contains the annotation. The default sheet name is
"Strain_annotations".