feat(strain.go): add LoadStrainFolderToTableFlag function to support folder input
The LoadStrainFolderToTableFlag function was added to allow the program
to accept a folder as input, which increases the flexibility of the
program as it can now process multiple files at once. The common flags
between LoadStrainToTableFlag and LoadStrainFolderToTableFlag were
extracted to a new function, strainToTableFlag, to avoid code
duplication and improve maintainability.