refactor(load.go): add omitempty to optional fields in PhenotypePayload struct
The 'omitempty' option indicates that the field should be omitted from
the encoding if the field has an empty value, defined as false, 0, a nil
pointer, a nil interface value, and any empty array, slice, map, or
string. This makes the JSON output cleaner when these fields are not
used.