feat(worker.go): add concurrent task processing functionality
A new file, worker.go, has been added to the concurrent package. This
file introduces a new feature that allows tasks to be processed
concurrently. The Task and TaskWrapper types have been defined to
represent a generic function and its input respectively. The ProcessWork
function takes a slice of TaskWrappers and processes them concurrently,
returning a slice of results and any errors that occurred.