feat(functional_handler.go): add new functional_handler.go file for phenotype package
This new file is added to handle the functionality related to phenotype
annotations. It includes functions for handling assayId, environmentId,
and phenotypeId, which are crucial for the phenotype annotation process.
The file also includes functions for processing ontology term Id and
marshalling payload, which are necessary for the proper functioning of
the phenotype annotation process.