feat(load.go): add phenotype loading functionality to PhenotypeLoader
The PhenotypeLoader now has the ability to add phenotype annotations and
phenotype rows. This is achieved by creating new instances of
PhenotypeLoader with the added annotations and then converting these
loaders to payloads. These payloads are then marshalled and sent as HTTP
requests to create new phenotype entries in the database. This
functionality is crucial for loading phenotype data into the
application.