feat(functional_handler.go): add assignedByIdHandler to handle assignedBy field
The assignedByIdHandler function has been added to handle the assignedBy
field in the PhenotypeLoader. This allows the assignedBy field to be
populated with the appropriate user ID when a non-zero aid is provided.
The assignedById function has been modified to search for the workspace
user using the workspace manager. This ensures that the correct user ID
is returned, or zero if the user is not found, improving the accuracy of
the assignedBy field.