refactor(cli): move common functions to functional_handlers.go for better code organization
The code has been refactored to move common functions to a separate
file, functional_handlers.go, for better code organization and
readability. This also makes the code more maintainable as changes to
these common functions will only need to be made in one place. New
functions for parsing and listing phenotype and strain files have been
added to enhance the functionality of the application.