docs(load.go): add comments to NewPhenotypeLoader and Load functions
The comments were added to provide more clarity on what the
NewPhenotypeLoader and Load functions do. The NewPhenotypeLoader comment
explains that it creates a new instance of PhenotypeLoader with the
provided properties. The Load comment explains that it processes the
phenotype data from the given reader and loads them into the database,
doing so concurrently up to a set limit before waiting and continuing
with the next batch.