feat(functional_handler.go): add ProcessEnvOntologyTerm function to handle specific string pattern
A new function, ProcessEnvOntologyTerm, has been added to handle a
specific string pattern in the input. This function checks if the input
string contains the character 'D' three or more times. If it does, it
removes the first occurrence of 'D'. This function was added to handle a
specific use case where the input string may contain 'D' multiple times,
and it's necessary to remove the first occurrence for further
processing.