feat(.github/workflows): add staging-build.yaml for automated staging builds
A new GitHub Actions workflow file, staging-build.yaml, has been added.
This workflow triggers a Docker build and push whenever there's a push
to the 'staging' branch. This automation simplifies the process of
deploying changes to the staging environment, ensuring that the staging
environment is always up-to-date with the latest changes.