fix(strainsyn.go): change condition to check if tac.Data is nil instead of checking its length
The condition to check if tac.Data is empty was changed from checking
the length of tac.Data to checking if tac.Data is nil. This change was
made because if tac.Data is nil, checking its length would cause a
runtime error. By checking if tac.Data is nil, we can avoid this
potential error.