fix(gene_updater.go): change the condition to check if all jobs are complete
The condition to check if all genes fetched from ArangoDB have been
processed is changed to check if all jobs have been submitted AND all
submitted jobs are complete. This is because the processing is now done
by a gRPC pool, and the progress is tracked by the number of jobs
submitted and completed.