fix(statement.go): reorder loops in GeneProductQuery to improve performance
The order of the loops in the GeneProductQuery has been changed.
Previously, the `gene_product` collection was iterated first, and then
the `locus_gp` collection was filtered based on the `gene_product_no`.
By reordering the loops to iterate over `locus_gp` first and then
`gene_product`, the query can potentially be more efficient by reducing
the number of `gene_product` documents that need to be examined. This
change aims to optimize the query execution time.